# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "WW API",
    "description": "Documentation automatically generated by the <b>swagger.autogen</b> module."
  },
  "host": "ww.lndo.site:80",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/product": {
      "post": {
        "tags": [
          "Product"
        ],
        "description": "Endpoint update a product",
        "parameters": [
          {
            "name": "id",
            "description": "productID",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Single product.",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Product"
        ],
        "description": "Endpoint add a new product",
        "parameters": [
          {
            "name": "id",
            "description": "productID",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Single product.",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Product"
        ],
        "description": "Endpoint delete a product",
        "parameters": [
          {
            "name": "id",
            "description": "productID",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/products/{catID}": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "Endpoint to get products by category ID",
        "parameters": [
          {
            "name": "catID",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "category ID"
          }
        ],
        "responses": {
          "200": {
            "description": "All products.",
            "schema": {
              "$ref": "#/definitions/Products"
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
    "Product": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 122850
        },
        "name": {
          "type": "string",
          "example": "WW by BUILT(r) 30 oz Tumbler"
        },
        "description": {
          "type": "string",
          "example": "WW by BUILT(r) 30 oz Tumbler 5052013P"
        },
        "url": {
          "type": "string",
          "example": "/ww-by-built-30oz-tumbler"
        },
        "isActive": {
          "type": "boolean",
          "example": true
        },
        "isNew": {
          "type": "boolean",
          "example": true
        },
        "skus": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "example": 123560
              },
              "name": {
                "type": "string",
                "example": "WW by BUILT(r) 30 oz Tumbler"
              },
              "description": {
                "type": "string",
                "example": "WW by BUILT(r) 30 oz Tumbler 5052013P"
              },
              "shortDescription": {
                "type": "string",
                "example": "Sip away! Now you can take your hot or cold drinks anywhere with this 30 oz. vacuum insulated tumbler with double walls--keeping drinks cold for 24 hours and hot for 6 hours."
              },
              "startDate": {
                "type": "string",
                "example": "2020-12-09T05:00:00.000+0000"
              },
              "retailPrice": {
                "type": "string",
                "example": "19.99"
              },
              "salePrice": {
                "type": "string",
                "example": "19.99"
              },
              "basePriceUnit": {
                "type": "string",
                "example": "g"
              },
              "discount": {
                "type": "string",
                "example": "0.00"
              },
              "isActive": {
                "type": "boolean",
                "example": true
              },
              "isAvailable": {
                "type": "boolean",
                "example": true
              },
              "inventoryType": {
                "type": "string",
                "example": "CHECK_QUANTITY"
              },
              "wwSku": {
                "type": "string",
                "example": "5052013"
              },
              "productType": {
                "type": "string",
                "example": "TangibleProduct"
              },
              "quantityAvailable": {
                "type": "number",
                "example": 471
              },
              "isSubscription": {
                "type": "boolean",
                "example": false
              },
              "taxCode": {
                "type": "string",
                "example": "P0000000"
              },
              "fulfillmentType": {
                "type": "object",
                "properties": {
                  "friendlyName": {
                    "type": "string",
                    "example": "Physical Ship"
                  },
                  "type": {
                    "type": "string",
                    "example": "PHYSICAL_SHIP"
                  }
                }
              },
              "media": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 124561
                    },
                    "url": {
                      "type": "string",
                      "example": "https://www.weightwatchers.com/us/shop/assets-proxy/weight-watchers/image/upload/q_auto/v1/prod/en-us-ec/static-asset/US_Built_Tumbler_52013-0266.jpg?auto=webp"
                    },
                    "altText": {
                      "type": "string",
                      "example": "WW by Built 30 oz Tumbler"
                    }
                  }
                }
              }
            }
          }
        },
        "medias": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "example": 124561
              },
              "url": {
                "type": "string",
                "example": "https://www.weightwatchers.com/us/shop/assets-proxy/weight-watchers/image/upload/q_auto/v1/prod/en-us-ec/static-asset/US_Built_Tumbler_52013-0266.jpg?auto=webp"
              },
              "altText": {
                "type": "string",
                "example": "WW by Built 30 oz Tumbler"
              }
            }
          }
        },
        "retailPrice": {
          "type": "string",
          "example": "19.99"
        },
        "salePrice": {
          "type": "string",
          "example": "19.99"
        },
        "discount": {
          "type": "string",
          "example": "0.00"
        },
        "currency": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string",
              "example": "USD"
            }
          }
        },
        "category": {
          "type": "object",
          "properties": {
            "id": {
              "type": "number",
              "example": 10372
            },
            "name": {
              "type": "string",
              "example": "Drinkware"
            },
            "active": {
              "type": "boolean",
              "example": true
            },
            "isFeatured": {
              "type": "boolean",
              "example": false
            },
            "navigationIncluded": {
              "type": "boolean",
              "example": true
            },
            "url": {
              "type": "string",
              "example": "/kitchen/drinkware"
            },
            "urlKey": {
              "type": "string",
              "example": "drinkware"
            },
            "parent": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 10188
                },
                "name": {
                  "type": "string",
                  "example": "Kitchen"
                },
                "url": {
                  "type": "string",
                  "example": "/kitchen"
                },
                "parent": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 10109
                    },
                    "name": {
                      "type": "string",
                      "example": "Kitchen & Lifestyle"
                    },
                    "url": {
                      "type": "string",
                      "example": "/kitchen-and-lifestyle"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "Products": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 42253
          },
          "name": {
            "type": "string",
            "example": "Triple Decker Chocolate Mini Bar"
          },
          "description": {
            "type": "string",
            "example": "Triple Decker Chocolate Mini Bar 6023017P"
          },
          "url": {
            "type": "string",
            "example": "/triple-decker-chocolate-mini-bar"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "isNew": {
            "type": "boolean",
            "example": true
          },
          "skus": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 42503
                },
                "name": {
                  "type": "string",
                  "example": "Triple Decker Chocolate Mini Bar"
                },
                "description": {
                  "type": "string",
                  "example": "Triple Decker Chocolate Mini Bar 6023017P"
                },
                "shortDescription": {
                  "type": "string",
                  "example": "Say hello to a mini bar with major flavor! The Triple Decker Chocolate Mini Bar combines fluffy marshmallow and flaky chocolate crisps covered in a dreamy chocolate layer. "
                },
                "startDate": {
                  "type": "string",
                  "example": "2020-07-29T04:00:00.000+0000"
                },
                "retailPrice": {
                  "type": "string",
                  "example": "7.99"
                },
                "salePrice": {
                  "type": "string",
                  "example": "7.99"
                },
                "basePriceUnit": {
                  "type": "string",
                  "example": "g"
                },
                "discount": {
                  "type": "string",
                  "example": "0.00"
                },
                "isActive": {
                  "type": "boolean",
                  "example": true
                },
                "isAvailable": {
                  "type": "boolean",
                  "example": true
                },
                "inventoryType": {
                  "type": "string",
                  "example": "CHECK_QUANTITY"
                },
                "wwSku": {
                  "type": "string",
                  "example": "6023017"
                },
                "productType": {
                  "type": "string",
                  "example": "TangibleProduct"
                },
                "quantityAvailable": {
                  "type": "number",
                  "example": 12117
                },
                "isSubscription": {
                  "type": "boolean",
                  "example": false
                },
                "taxCode": {
                  "type": "string",
                  "example": "PF050900"
                },
                "fulfillmentType": {
                  "type": "object",
                  "properties": {
                    "friendlyName": {
                      "type": "string",
                      "example": "Physical Ship"
                    },
                    "type": {
                      "type": "string",
                      "example": "PHYSICAL_SHIP"
                    }
                  }
                },
                "media": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 107001
                      },
                      "url": {
                        "type": "string",
                        "example": "https://www.weightwatchers.com/us/shop/assets-proxy/weight-watchers/image/upload/q_auto/v1/prod/en-us-ec/static-asset/6023017_TripleDeckerChoc_Box_f.jpg?auto=webp"
                      },
                      "altText": {
                        "type": "string",
                        "example": "Triple Decker Mini Bar"
                      }
                    }
                  }
                }
              }
            }
          },
          "medias": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 107001
                },
                "url": {
                  "type": "string",
                  "example": "https://www.weightwatchers.com/us/shop/assets-proxy/weight-watchers/image/upload/q_auto/v1/prod/en-us-ec/static-asset/6023017_TripleDeckerChoc_Box_f.jpg?auto=webp"
                },
                "altText": {
                  "type": "string",
                  "example": "Triple Decker Mini Bar"
                }
              }
            }
          },
          "retailPrice": {
            "type": "string",
            "example": "7.99"
          },
          "salePrice": {
            "type": "string",
            "example": "7.99"
          },
          "discount": {
            "type": "string",
            "example": "0.00"
          },
          "currency": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "USD"
              }
            }
          },
          "category": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "example": 10131
              },
              "name": {
                "type": "string",
                "example": "Snack bars"
              },
              "active": {
                "type": "boolean",
                "example": true
              },
              "isFeatured": {
                "type": "boolean",
                "example": false
              },
              "navigationIncluded": {
                "type": "boolean",
                "example": true
              },
              "url": {
                "type": "string",
                "example": "/food/snack-bars"
              },
              "urlKey": {
                "type": "string",
                "example": "snack-bars"
              },
              "parent": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 10111
                  },
                  "name": {
                    "type": "string",
                    "example": "Food & Drink"
                  },
                  "url": {
                    "type": "string",
                    "example": "/food"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}